consumer/requirements.txt
faust-streaming
httpx